Linked Documents
Fixed issue in resolvers, now the User docu and the Room docu are linked together correctly. Currently the way things seed, each user and room are paired up one-to-one, this works fine for our case because it ensures no users get assigned multiple times so long as there are more users than rooms.